requirements.txt
click==8.0.4
cycler==0.11.0
Flask==2.0.3
fonttools==4.30.0
gunicorn==20.1.0
imageio==2.16.1
itsdangerous==2.1.1
Jinja2==3.0.3
jsonpickle==2.1.0
kiwisolver==1.3.2
MarkupSafe==2.1.0
matplotlib==3.5.1
networkx==2.7.1
numpy==1.22.3
opencv-python==4.5.5.64
packaging==21.3
Pillow==9.0.1
pyparsing==3.0.7
python-dateutil==2.8.2
PyWavelets==1.2.0
scikit-image==0.19.2
scipy==1.8.0
six==1.16.0
tifffile==2022.2.9
torch==1.12.1
torchvision==0.11.3
typing-extensions==4.1.1
Werkzeug==2.0.3
